DiscountCondition를 인터페이스로 변경하고 각 타입을 분리
Period, Sequence 기반으로 구현해 사용하도록 변경
