+++ src/jooom/chapter05/DiscountCondition.java
@@ -1,29 +1,5 @@
 package jooom.chapter05;
 
-import java.time.DayOfWeek;
-import java.time.LocalTime;
-
-public class DiscountCondition {
-    private DiscountConditionType type;
-    private int sequence;
-    private DayOfWeek dayOfWeek;
-    private LocalTime startTime;
-    private LocalTime endTime;
-
-    public boolean isSatisfiedBy(Screening screening){
-        if (type == DiscountConditionType.PERIOD){
-            return isSatisfiedByPeriod(screening);
-        }
-
-        return isSatisfiedBySequence(screening);
-    }
-
-    private boolean isSatisfiedByPeriod(Screening screening){
-        return dayOfWeek.equals(screening.getWhenScreened().getDayOfWeek()) &&
-                startTime.isBefore(screening.getWhenScreened().toLocalTime()) &&
-                endTime.isAfter(screening.getWhenScreened().toLocalTime());
-    }
-    private boolean isSatisfiedBySequence(Screening screening){
-        return sequence == screening.getSequence();
-    }
+public interface DiscountCondition {
+    boolean isSatisfiedBy(Screening screening);
 }
